requirements.txt
hachoir==3.3.0
pillow==11.0.0
pillow_heif==0.20.0
pyexiv2==2.15.3
